+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Logo changed
 - Unsaved changes require confirmation to be discarded
 - Shortcut: Use `CTRL+S` to save files
+- Shortcut: Use `CTRL+SHIFT+S` to save as...
+- Shortcut: Use `CTRL+O` to open files
+- Shortcut: Use `CTRL+N` to make new files
 
 ### Fixed
 - File extension bug while overwriting file